More optimizations
Moved various small chunks of code out of ReaderSessionsChart to keep it and dependencies out of initial load. Restrict fetched collection fields to the ones we need.
