+++ package.json
@@ -16,7 +16,8 @@
         "build-storybook": "build-storybook -s public",
         "corsChrome": "chrome http://localhost:3000/ --disable-web-security --auto-open-devtools-for-tabs --user-data-dir=%temp%/chromeTemp",
         "crowdin-dangerous-upload": "ts-node --compiler-options \"{\\\"module\\\": \\\"commonjs\\\"}\" src/localization/crowdin-sync -- upload",
-        "crowdin-download": "ts-node --compiler-options \"{\\\"module\\\": \\\"commonjs\\\"}\" src/localization/crowdin-sync -- download"
+        "crowdin-download": "ts-node --compiler-options \"{\\\"module\\\": \\\"commonjs\\\"}\" src/localization/crowdin-sync -- download",
+        "analyze": "source-map-explorer build/static/js/*chunk.*"
     },
     "dependencies": {
         "@contentful/rich-text-react-renderer": "^13.4.0",
@@ -95,9 +96,9 @@
         "@types/react-lazyload": "^2.6.0",
         "@types/react-router-dom": "^5.1.5",
         "@types/react-select": "^3.0.11",
-        "@types/swiper": "^5.4.1",
         "@types/storybook__addon-info": "^4.1.2",
         "@types/storybook__react": "^4.0.2",
+        "@types/swiper": "^5.4.1",
         "async-retry": "^1.3.1",
         "babel-loader": "^8.0.6",
         "bloom-player": "^2.0.33",
@@ -107,6 +108,7 @@
         "fs-extra": "^9.0.1",
         "ncp": "^2.0.0",
         "react-select": "^3.0.8",
+        "source-map-explorer": "^2.5.2",
         "temp-dir": "^2.0.0",
         "ts-jest": "^25.2.0",
         "ts-loader": "^6.2.1",
